Extract audio engine option helpers
